requirements.txt
customtkinter
playwright
pyperclip
Pillow